+++ src/storage/store/store_async/opendal.rs
@@ -5,8 +5,8 @@ use crate::{
     byte_range::{ByteRange, InvalidByteRangeError},
     storage::{
         AsyncListableStorageTraits, AsyncReadableStorageTraits, AsyncReadableWritableStorageTraits,
-        AsyncWritableStorageTraits, StorageError, StoreKey, StoreKeyRange, StoreKeyStartValue,
-        StoreKeys, StoreKeysPrefixes, StorePrefix,
+        AsyncWritableStorageTraits, StorageError, StoreKey, StoreKeyStartValue, StoreKeys,
+        StoreKeysPrefixes, StorePrefix,
     },
 };
 
@@ -87,13 +87,6 @@ impl AsyncReadableStorageTraits for AsyncOpendalStore {
         } else {
             Ok(None)
         }
-    }
-
-    async fn get_partial_values(
-        &self,
-        key_ranges: &[StoreKeyRange],
-    ) -> Result<Vec<MaybeBytes>, StorageError> {
-        self.get_partial_values_batched_by_key(key_ranges).await
     }
 
     async fn size_prefix(&self, prefix: &StorePrefix) -> Result<u64, StorageError> {